refactor!: remove `app.main` module and instead use `app` module directly
